initial stab at bulk binding.
This is really bad code, needs a lot of cleaning up. But it was pretty
easy to write, and I think that with a few more functions or classes it
should be tottally not a big deal. B'kitzur, its a humble beggining. As
far of proof-of-concept is concerned, I think it shows that I'm going in
the right direction in terms of creating accesible and manipulatible
data.